hotfix

+++ nts.uk/uk.com/com.ctx/bs/employee/nts.uk.ctx.bs.employee.ws/src/main/java/nts/uk/ctx/bs/employee/ws/jobtitle/sequence/SequenceMasterWebService.java
@@ -39,7 +39,7 @@ public class SequenceMasterWebService extends WebService {
 	 */
 	@Path("findMaxOrder")
 	@POST
-	public short findMaxOrder() {
+	public int findMaxOrder() {
 		return this.sequenceMasterFinder.findMaxOrder();
 	}
